backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
boto3==1.34.0
chromadb==0.4.18
sentence-transformers==2.2.2
huggingface-hub==0.16.4
langchain-community>=0.0.2
pydantic==2.5.0
python-dotenv==1.0.0 